code review: better abstraction + comments
